requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
PyJWT
python-multipart
pydantic[email-validator]
aiofiles
pg8000
